Move k8s client within common folder
The k8s client can be used by different servers
