+++ config/rbac/role.yaml
@@ -9,6 +9,11 @@ rules:
   verbs:
   - create
   - post
+- nonResourceURLs:
+  - /o2ims-infrastructureInventory/v1/resourceTypes
+  verbs:
+  - get
+  - list
 - apiGroups:
   - ""
   resources:
