+++ internal/service/common/clients/constants.go
@@ -1,0 +1,8 @@
+package clients
+
+import "time"
+
+const (
+	ListRequestTimeout   = 30 * time.Second
+	SingleRequestTimeout = 10 * time.Second
+)
